Add join_strings() call and use it from mattrs()
This is a super-nit kind of fix, but the fact that we had rerolled a join-strings algo in a half-dozen places made my teeth hurt, so I decided to fix it:

- Add join_strings() to Util.h
- revise the mattrs() calls to use it instead of the janky mess they used

This doesn't move the needle on code size or speed but it is less weird.

Probably other places we could/should use this too.

(Does C++20 have join/split strings in the std library yet? If not, why not?)