Bump serialization version to 18.0.0
As a matter of policy, we should probably bump the version of the serialization format for every version of Halide -- even if changes are minimal-to-nonexistent -- to reinforce the fact that this isn't intended in any way as a long-term archival format.

This PR suggests that we bump the major version to match the main Halide version, but I'm open for other suggestions.