Fix an issue where the Halide compiler hits an internal error when bool types are used with e.g. widening_mul. This situation did not arise from user code doing this directly, but rather through some chain o lowering with float16 types. The test cases added to correctness_intrinsics target the issue directly and do fail without the fix.
I did not add broader coverage for bool types and intrinsics as it
would require more thinking. Most of them overflow for the true/true
case and thus are of questionable use, however widening operations
cannot overflow... Certainly we could define the language to forbid
this, but currently the frontend does not do so. As indicated above,
the use case driving this was not using bool arithmetic to begin with.