Handle loads of broadcasts in FlattenNestedRamps (#8139)
With sufficiently perverse schedules, it's possible to end up with a
load of a broadcast index (rather than a broadcast of a scalar load).
This made FlattenNestedRamps divide by zero. Unfortunately this happened
in a complex production pipeline, so I'm not entirely sure how to
reproduce it. For that pipeline, this change fixes it and produces
correct output.