A 'broadcast' shuffle is more complex than it seems
I was poking at the Shuffle node, and checking its usage, and it seems
that despite the comment, Shuffles that return true for is_broadcast are
not the same as a Broadcast node. Instead of repeating the input vector
some number of times, it repeats a shuffle of the input vector. This
means IRPrinter was incorrect. None of the other usages were bad.

This PR makes this clearer in the comment, and fixes IRPrinter.