Add a simple CSV-based image format to halide_image_io
This is intended to be used with RunGen to facilitate passing small image data on the commandline when desirable or necessary. It pains me to create a bespoke image format here, but there doesn't appear to be a well-known image format that fills this need:
- plaintext
- capable of representing ~all the numeric types that Halide can handle (though f16/bf16 are TODO in this PR)
- capable of representing arbitrary dimensionality
- doesn't require a complex library dependency to read (e.g. libTIFF)

I plan on making use of this format in some upcoming improvements to RunGen and the Visualizer code.