More aggressively unify duplicate lets (#8204)
* Make unify_duplicate_lets more aggressive

The simplifier can also clean up most of these, but it's harder for it
because it has to consider that other mutations may have taken place.
Beefing this up has no impact on lowering times for most apps, but
something pathological was going on for local_laplacian. At 20 pyramid
levels, this speeds up lowering by 1.3x. At 50 pyramid levels it's 2.3x.
At 100 pyramid levels it's 4.1x.

It also slightly reduces binary size.

* Clarify comment; Avoid double-lookup into the scope

Looking up with an Expr key and deep equality is expensive, so this was
bad.

* Add a std::move