Fix associative op detection for saturating_add
The old pattern was min(u16(x8 + y8), any u16 constant at all), which is
wrong in several ways. I replaced that with all the variants of
saturating_add I could think of.

It would be nice to be able to call find_intrinsics here to normalize,
but that would require making find_intrinsics trigger on scalars, and I
wanted to make the minimal change here to fix things.