Quick CMake fixes enabled by 3.28 (#8365)
* Use FindCUDAToolkit in apps/cuda_mat_mul

* Replace dummy FindHalide.cmake with pkg redirects

Having the dummy file on disk is confusing and is easy
to accidentally install when modifying CMake install
rules. Better to use the CMake 3.24+ feature of the
package redirects dir to truly disable find_package
for Halide inside the build.

* Avoid creating a dummy file for Halide_Python

* Fix formatting in CMakeLists.txt

* Add SpirvIR.h to the list of Halide sources

* Use BUILD_LOCAL_INTERFACE in PyStubs

* Consistently use HALIDE_H variable

* Comment overriding POSITION_INDEPENDENT_CODE

* Check Halide_STATIC_DEFINE at configure time.

This avoids sending a generator expression downstream. These
are functionally identical, but it's just one less thing to
evaluate.

* Use BUILD_LOCAL_INTERFACE for SPIRV-Headers