Scan generated export files to determine dependencies. (#8385)
This commit contains a module for declaring that an export file might
depend on another CMake package that was found by find_package. Such
dependencies are collected in a project-wide property (rather than a
variable) along with a snippet of code that reconstructs the original
call.

Then, after we have installed an export file via install(EXPORT), we can
call a helper to add install rules that will read the file as-generated
by CMake to check whether any of these packages could be required.

CMake does not like to expose this information, in part because
generator expressions make computing the eventual link set undecidable.
Even so, for our purposes if Pkg:: appears in our link-libraries list,
then we need to find_package(Pkg).

This module implements that heuristic.

So why is this hard? It's because checking whether a dependency is
actually included is very complicated. A library will appear if:

1. It is SHARED or MODULE
2. It linked privately to a STATIC target
    - These appear as $<LINK_ONLY:${dep}>
3. It is STATIC and linked publicly to a SHARED target;
4. It is INTERFACE or ALIAS and linked publicly
5. It is included transitively via (4) and meets (1), (2), or (3)
6. I am not sure this set of rules is exhaustive.

There is an experimental feature in CMake 3.30 that will some day
replace this module.