Rewrite bundle_static to be much more efficient. (#8386)
The `bundle_static` function now detects the private static dependencies
on the given target (in our case, always Halide) and uses the platform
librarian tool to merge static dependencies into a static library. It
picks which tool to use by checking, in order:

* When targeting Windows, it looks for `lib.exe`.
* When targeting macOS, it checks if `libtool` is the Apple libtool.
* Whether `ar` is GNU ar and if so, generates an MRI script.
* Otherwise, a `FATAL_ERROR` is issued.

To mark a static library for bundling, we link privately and use
the `$<BUILD_LOCAL_INTERFACE:...>` generator expression. This prevents
it from being exported, too.

The generator expression that implements this logic is quite complex.
It involves meta-programming generator expressions during evaluation
and then evaluating them. Even so, this saves a considerable amount of
time unpacking LLVM into a temporary directory and adding the objects
to the link line (the previous approach).