Update README.md
LLVM recently started defaulting ENABLE_ZSTD to on, but on macos brew puts libzstd into a directory that's not in the linker search path, and llvm-config --system-libs declares a dependency on zstd but doesn't add the correct search path.  It's not clear why we need llvm with zstd, so I'm proposing just disabling it in our build instructions to avoid this failure mode.