Include our Markdown documentation in the Doxygen site.
A few quirks in the Markdown parser were worked around here.

The most notable is that the sequence `]:` causes Doxygen
to interpret a would-be link as a trailing reference even
if it is not at the start of a line. Duplicating the single
bracket reference is a portable workaround, i.e.

    [winget]  ~>  [winget][winget]

It also doesn't stop interpreting `@` directives inside inline
code, so it warns about our use of the `@` as a decorator symbol
inside Python.md.