Select condition vector lanes must match the true and false value (#8465)
The ability for the condition to be a scalar while the value is a vector
was a mistake that added complexity in a few places. It's easy enough to
just check if the condition is a broadcast if you really need to know.

One place it added a bunch of complexity is that it meant in the RHS of
the simplifier rules, sometimes you needed an implicit broadcast. This
was responsible for about a third of the code size in Simplify_Sub.o!
It's also unclear if it respects the reduction order we use to prove the
simplifier terminates, because those rules were turning an implicit
broadcast in the IR into a new actual Broadcast node.