Handle some misc TODOs
CodeGen_LLVM.cpp: opaque pointers are now standard, and that flag no
longer works.

Var.h: We convert strings to Vars in many places internally, and some of
those Vars originated from implicit Vars, so it's not feasible to
require than the version that takes an explicit string isn't allowed to
be passed things of the form "_[0-9]*". You can use the explicit
constructor to make collisions with other Vars, and yes this includes
the implicit vars.