Rewrite the rfactor scheduling directive (#8490)
The old implementation suffered from several serious issues. It
duplicated substantial amounts of the logic in ApplySplit.cpp, the
way it handled adapting the predicate to the reducing func was
unprincipled, and it confused dims and vars in a way that could
segfault. It also left the order of pure dimensions unspecified.
The new implementation chooses to follow the existing dims list.

We now disallow rfactor() on funcs with RVar+Var fused schedules.
The implementation also relies on a new or_condition_over_domain
helper function and drops the purify() scheduling directive.

Fixes #7854