Propagate info in more places
Some fuzzing with extra checks turned on revealed that the simplifier
would often produce constants without the ExprInfo knowing they were
constants. In a few places we were doing a janky remutation of a known
constant just to set the info, but this wasn't being consistently done.
This commit changes constant-folding in the simplifier and info handling
in boolean ops to know that it's a constant value being returned.

Also standardize the spelling of zeros