Move to opaque llvm pointers
For a long time llvm has been slowly moving to a model where pointers
are an opaque type, instead of knowing the pointee type. This has been
complete for a while now, so they've started deprecating the APIs for
constructing a pointer with a pointee type (these APIs ignore the
pointee type anyway), or doing pointer casts, which are no-ops. This PR
removes use of those APIs and is a nice simplification of some code.