Drop check for LLVM_LIBCXX in FindHalide_LLVM.cmake
This was never tested in CI, and grepping the LLVM sources (18 - main)
for this variable comes up empty. Mixing standard libraries in a single
project isn't a well-supported scenario anyway, so users wishing to
override the standard library should use a consistent toolchain file
between LLVM and Halide to do so.