Add filtering capabilities to HL_DEBUG_CODEGEN (#8627)
This PR extends the semantics of HL_DEBUG_CODEGEN with features for
filtering messages only from certain files, functions, and lines.

The new syntax for HL_DEBUG_CODEGEN is a ;-delimited list of rules
of the form:

    verbosity[,file_suffix[:line_low[-line_high]]][@function_suffix]

* verbosity is a positive integer indicating the maximum (inclusive)
  verbosity level to emit. This is equivalent to the old semantics.
* file_suffix is a maximal string of non-:@ characters. If it is
  specified, the rule applies only to files ending in this string.
  Typically, this will be a path fragment like src/Func.cpp.
* The values line_low and line_high specify a range of line numbers,
  inclusive on both sides. If neither is specified, they are 0 and
  INT_MAX. If only line_low is specified, line_high is set equal to
  line_low.
* function_suffix is a string. If it is specified, the rule applies
  only to functions (as determined by __FUNCTION__) whose names end in
  this string. Typically, this will be the whole name of a function
  (e.g. rfactor).

The motivation is to enable higher-signal print debugging. For example,
a useful setting while debugging rfactor might be:

    1;5,Func.cpp@rfactor;5,Associativity.cpp

This commit also replaces Halide::Internal::debug with a macro that is
no longer exported by default. To regain access to the internal Halide
debugging and error handling macros, define HALIDE_KEEP_MACROS before
including Halide.h

Thanks to Martijn Courteaux for suggesting the syntax for the rules.

---------

Co-authored-by: Martijn Courteaux <courteauxmartijn@gmail.com>