Constrain Clang and LLD searches to LLVM version
As of LLVM 17 or so, the *ConfigVersion.cmake files have been correctly
shipped with Clang and LLD, not just LLVM. The FindHalide_LLVM.cmake
module now searches for the exact matching versions of Clang and LLD
using the LLVM_VERSION variable. Note that the LLVM_PACKAGE_VERSION
variable is not suitable for this purpose as it might contain a git
hash suffix.

Fixes #8633